+++ package.json
@@ -31,7 +31,7 @@
     "react-native-modal": "^4.1.1",
     "react-native-navigation": "^1.1.328",
     "react-native-tab-view": "^0.0.74",
-    "react-native-ui-lib": "^2.0.156",
+    "react-native-ui-lib": "^2.0.158",
     "react-native-vector-icons": "^4.5.0"
   },
   "devDependencies": {
