+++ src/modules/auth/screens/SignIn/index.tsx
@@ -4,21 +4,21 @@ import { View, Image, StyleSheet, ViewStyle, Text, TextStyle } from 'react-nativ
 import ScreenWrapper from '../../../common/components/ScreenWrapper'
 
 export default class SignInScreen extends React.Component {
-  static navigatorStyle = {
-    navBarHidden: true
+  public static navigatorStyle = {
+    navBarHidden: true,
   }
-  handleSubmit = (values: IValues) => {
+  public handleSubmit = (values: IValues) => {
     return new Promise(() => {
       alert(JSON.stringify(values))
     })
   }
-  render() {
+  public render() {
     return (
       <ScreenWrapper>
         <Image
           source={{uri: 'https://i.pinimg.com/originals/06/68/99/066899986813f160faf1a641c4b429a1.png'}}
           style={styles.logo}
-          resizeMode="contain"
+          resizeMode='contain'
         />
         <Text style={styles.title}>
           Quack Pack
@@ -29,9 +29,9 @@ export default class SignInScreen extends React.Component {
             validate={() => null}
           />
         </View>
-        
+
       </ScreenWrapper>
-      
+
     )
   }
 }
@@ -42,15 +42,15 @@ const styles = StyleSheet.create({
     marginBottom: 10,
     width: 200,
     height: 200,
-    alignSelf: 'center'
+    alignSelf: 'center',
   } as ViewStyle,
   title: {
     fontSize: 30,
     fontWeight: '900',
-    textAlign: 'center'
+    textAlign: 'center',
   } as TextStyle,
   formWrapper: {
     flex: 1,
     marginTop: 40,
-  } as ViewStyle
+  } as ViewStyle,
 })
