+++ src/modules/delivery/containers/ConfirmDelivery/index.tsx
@@ -1,0 +1,11 @@
+import ConfirmDelivery from '../../components/ConfirmDelivery'
+
+export default () => <ConfirmDelivery
+  isVisible
+  onRequestClose={() => {
+    alert('request close')
+  }}
+  onQRCodeRead={() => {
+    alert('qrcode read')
+  }}
+/>
