+++ src/modules/delivery/containers/CurrentDelivery/index.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react'
+import CurrentDelivery from '../../components/CurrentDelivery'
+
+export default () => (
+  <CurrentDelivery />
+)
