+++ src/modules/delivery/containers/DeliverySuccess/index.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react'
+import DeliverySuccess from '../../components/DeliverySuccess'
+
+export default () => <DeliverySuccess
+  onPressFinish={() => {
+    alert('finish')
+  }}
+/>
