+++ src/modules/delivery/containers/WaitingDelivery/index.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react'
+import WaitingDelivery from '../../components/WaitingDelivery'
+
+export default () => (
+  <WaitingDelivery />
+)
