+++ src/modules/delivery/screens/AcceptDelivery/index.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import ScreenWrapper from '../../../common/components/ScreenWrapper'
+import AcceptDelivery from '../../containers/AcceptDelivery'
+
+export default class AcceptDeliveryScreen extends React.Component {
+  public static navigatorStyle = {
+    navBarHidden: true,
+  }
+  public render() {
+    return (
+      <ScreenWrapper>
+        <AcceptDelivery />
+      </ScreenWrapper>
+    )
+  }
+}
