+++ src/modules/delivery/screens/CurrentDelivery/index.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import ScreenWrapper from '../../../common/components/ScreenWrapper'
+import CurrentDelivery from '../../containers/CurrentDelivery'
+
+export default class CurrentDeliveryScreen extends React.Component {
+  public static navigatorStyle = {
+    navBarHidden: true,
+  }
+  public render() {
+    return (
+      <ScreenWrapper>
+        <CurrentDelivery />
+      </ScreenWrapper>
+    )
+  }
+}
