+++ src/modules/location/components/LocationsListItem/index.tsx
