--- src/modules/order/containers/Locations/index.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react'
-import LocationsList from '../../../common/components/LocationsList'
-
-const locations = [{
-  id: 'a',
-  name: 'Mc Donalds',
-  address: 'R. caralhito',
-}, {
-  id: 'b',
-  name: 'Jorge Risco',
-  address: 'R. Cantareira, 123',
-}]
-
-export default () => <LocationsList locations={locations} />
