+++ src/navigation.tsx
@@ -1,0 +1,11 @@
+import { Navigation } from 'react-native-navigation'
+import { Screens } from './resources/screens'
+
+export const navigateToSignIn = () => {
+  Navigation.startSingleScreenApp({
+    screen: {
+      screen: Screens.DeliverySuccess,
+    },
+    animationType: 'fade',
+  });
+}
