+++ src/resources/animations/loading.ts
