+++ src/resources/colors.ts
@@ -1,4 +1,4 @@
 export const Colors = {
   PRIMARY_COLOR: '#FF9100',
-  WHITE: '#FFF'
+  WHITE: '#FFF',
 }
